% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aps_sco_synth.R
\docType{data}
\name{aps_sco_synth}
\alias{aps_sco_synth}
\title{Synthetic survey of Scottish labour force participation and ethnicity}
\format{
A tbl_df with with 26,137 rows and 2 variables:
\describe{
\item{ethnicity}{identifies as belonging to white or minority (non-white) ethnic group}
\item{unemployed}{identifies as unemployed, i.e., someone (1) without a job,
(2) who can start work within two weeks, and (3) who had either looked for work in
the previous four weeks or were waiting to start a job}
}
}
\source{
Synthetic data created by Eric Hall \email{ehall001@dundee.ac.uk},
University of Dundee, 2022.08.18.
}
\usage{
aps_sco_synth
}
\description{
This data set is for teaching and training purposes.
A data set of fictitious results from a survey of Scottish labour force participation and ethnicity.
The Scottish labour force refers to people in Scotland aged 16 and older who
are economically active. Unemployment here only relates to people in the labour
force, i.e., who are "economically active". Unemployment refers to people
without a job, who can start work within two weeks following their survey
interview, and who had either looked for work in the previous four weeks or
were waiting to start a job.
}
\examples{
data(aps_sco_synth)
}
\keyword{Scotland,}
\keyword{data}
\keyword{datasets,}
\keyword{ethnicity,}
\keyword{synthetic}
\keyword{unemployment,}
